Removing old integration tests
I moved to more unit testing.  I want to still break apart
the interface to a more generic responsibility of
1) Build a request
2) Generic "html" for Request.Do()
3) Unpack a response

This will allow for very strong unit testing around request building,
and response parsing.  We shouldn't be testing if we can post or receive
via http, just that our interfaces do what they are supposed to with a
given request or given response.

	deleted:    customer_integration_test.go
	deleted:    plan_integration_test.go
	deleted:    subscription_integration_test.go
	deleted:    transaction_integration_test.go