Start of a new take on the interface and errors
There is a lot of work here to do yet, but I'm
starting to flush out a pattern that involves
just the initial entity (i.e. card) and from
there you just call your methods (i.e. create)

Objects will now have an "Errors" struct as well
so that if we get api errors back, the can be consumed
directly from the object.  We can create convenience
helpers around them if needed as well.

I'm not happy with the "Card.Create" method yet.  Still
needs some more abstraction, but the building blocks are there
now.

The big win here is having a generic "http.DefaultClient.Do()"
as well as a generic "unwrap" function for the gzipped reaponse.

If I get down the patter for these correctly, it will allow me to unit
test the Card objects directly with XML.  This will allow us to test
a lot of scenarios quickly and effieciently.

So, to sum it up, lot of work yet, but headed in the right direction
I think.

	modified:   credit_card.go
	new file:   credit_card_errors.go
	new file:   credit_card_errors_test.go
	modified:   credit_card_test.go
	new file:   request.go
	new file:   xml_request.go